feat: add type for put and delete operation, change SN from integer to long, rewrite tests

+++ src/main/java/az/zeynalov/memtable/Header.java
@@ -6,6 +6,6 @@ import java.util.Objects;
 public record Header(
    int keySize,
    MemorySegment key,
-   int SN,
+   long SN,
    byte type){
 }
